Allow farmers to login/signup on frontend
Allows users to sign up for a new account from the front end, or simply
store a Basic Auth token to authorize later requests.

Enabled by a bunch of little services and components, some of which will
be expanded later.

Bit of a hack in many ways, and for the sake of a quick demo, I perhaps
should have skipped a multi-user implementation, but what is the point
of shrimp farms if you can't track your own?