Fix resources being created twice
A particularly fun bug, I perhaps got too clever for my own good.
The Normalize db query is meant to accept a resource, or and expression
that produces a resource. However, the way it was written, it used the
parameter twice, meaning if that parameter was an expression, the
expression *would be written twice in the final DB query*.

Whoops.

Solved by using a Let to ensure that the expression itself only
appears once in the query.