Replace farm size attribute with live calculation
Although it might be a bit faster to calculate farm size on
pond changes, it means multiple sources of truth and all the
potential for bugs that entails. A live calculation each time
farms are fetched should be fine in terms of resource usage
and is a considerably simpler solution.